feat: Expose list of tests with less magic
Prior to this change the static variable `TESTS` had to be declared in
the root of the crate in which the `#[libtest2::test]` macros were used.
This was automatically inserted by the `#[libtest2::main]` macro, but
if the user did not include this an error would be shown.

A consequence of this is that the `#[libtest2::test]` macro could not be
used on its own to collect tests, unless the user accessed the hidden,
private API exposed in the `_private` module.

The obvious downside of this change is that the user will no longer
receive an error message if they use the `#[libtest2::test]` macro
without `#[libtest2::main]`, and thus no tests will be run. The compiler
will, however, still complain if the user is missing a `main` function.